change receiver type from Int to Long

+++ sphinx/application/network/concepts/queries/concept-network-query-message/src/main/java/chat/sphinx/concept_network_query_message/model/MessageDto.kt
@@ -10,7 +10,7 @@ data class MessageDto(
     val chat_id: Long?,
     val type: Int,
     val sender: Long,
-    val receiver: Int?,
+    val receiver: Long?,
     val amount: Long,
     val amount_msat: Long,
     val payment_hash: String?,
